refactor: move the logic of handling matching event to matcher

+++ cmd/worker/matching_engine/config.go
@@ -6,7 +6,7 @@ type Config struct {
 	App   App   `envPrefix:"APP_"`
 	Kafka Kafka `envPrefix:"KAFKA_"`
 
-	TickNum int `env:"TICK_NUM" envDefault:"5"`
+	TickNum int8 `env:"TICK_NUM" envDefault:"5"`
 }
 
 type App struct {
